Patch for SF bug #2016130 checksum checker can not retrieve very large bitstream
git-svn-id: http://scm.dspace.org/svn/repo/branches/dspace-1_5_x@3332 9c30dcfa-912a-0410-8fc2-9e0234be79fd

+++ dspace-api/src/main/java/org/dspace/checker/BitstreamInfo.java
@@ -141,7 +141,7 @@ public final class BitstreamInfo
      * @param procStartDate
      *            When the last bitstream check started.
      */
-    public BitstreamInfo(boolean del, int storeNo, int sz, String bitstrmFmt,
+    public BitstreamInfo(boolean del, int storeNo, long sz, String bitstrmFmt,
             int bitstrmId, String usrFmtDesc, String intrnlId, String src,
             String chksumAlgorthm, String chksum, String nm, Date procEndDate,
             boolean toBeProc, Date procStartDate)
@@ -204,7 +204,7 @@ public final class BitstreamInfo
      * 
      * @return int
      */
-    public int getSize()
+    public long getSize()
     {
         return dspaceBitstream.getSize();
     }
